+++ beacon/staking/logs/types.go
@@ -35,6 +35,5 @@ import (
 // StakingService is the interface for the staking service.
 type StakingService interface {
 	AcceptDeposit(ctx context.Context, deposit *consensusv1.Deposit) error
-	ApplyDeposits(ctx context.Context) error
 	ProcessWithdrawal(ctx context.Context, withdrawal *enginev1.Withdrawal) error
 }
